Refactor(plugins): Move jinja filter code for arista.avd.generate_route_target to PyAVD
