+++ ansible_collections/arista/avd/plugins/filter/generate_route_target.py
@@ -5,7 +5,21 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
-import re
+from ansible.errors import AnsibleFilterError
+
+from ansible_collections.arista.avd.plugins.plugin_utils.pyavd_wrappers import RaiseOnUse, wrap_filter
+
+PLUGIN_NAME = "arista.avd.convert_dicts"
+
+try:
+    from pyavd.j2filters.convert_dicts import generate_route_target
+except ImportError as e:
+    generate_route_target = RaiseOnUse(
+        AnsibleFilterError(
+            f"The '{PLUGIN_NAME}' plugin requires the 'pyavd' Python library. Got import error",
+            orig_exc=e,
+        )
+    )
 
 DOCUMENTATION = r"""
 ---
@@ -36,30 +50,8 @@ _value:
 """
 
 
-def generate_route_target(esi_short):
-    """
-    generate_route_target Transform 3 octets ESI like 0303:0202:0101 to route-target
-
-    Parameters
-    ----------
-    esi : str
-        Short ESI value as per AVD definition in eos_designs
-
-    Returns
-    -------
-    str
-        String based on route-target format like 03:03:02:02:01:01
-    """
-    if esi_short is None:
-        return None
-    delimiter = ":"
-    esi = esi_short.replace(delimiter, "")
-    esi_split = re.findall("..", esi)
-    return delimiter.join(esi_split)
-
-
 class FilterModule(object):
     def filters(self):
         return {
-            "generate_route_target": generate_route_target,
+            "generate_route_target": wrap_filter(PLUGIN_NAME)(generate_route_target),
         }
