+++ python-avd/pyavd/j2filters/generate_route_target.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+import re
+
+def generate_route_target(esi_short: str | None) -> str:
+    """
+    generate_route_target Transform 3 octets ESI like 0303:0202:0101 to route-target
+
+    Parameters
+    ----------
+    esi : str
+        Short ESI value as per AVD definition in eos_designs
+
+    Returns
+    -------
+    str
+        String based on route-target format like 03:03:02:02:01:01
+    """
+    if esi_short is None:
+        return None
+    delimiter = ":"
+    esi = esi_short.replace(delimiter, "")
+    esi_split = re.findall("..", esi)
+    return delimiter.join(esi_split)
